Make SortableColumn and Search reusable
